Changed new ticket page
The new ticket page is now an almost fully working POST form. Data is sent to the server, which you will see the form of when you press submit.

The page also receives hard-coded information from the route 'newticket' in web.php (name, surname, ID etc. ) which can later on be substituted with values from the database